refactor pydantic models
