+++ backend/models/__init__.py
@@ -1,0 +1,23 @@
+"""
+Pydantic models for request/response validation.
+"""
+from .bridge import (
+    BridgeTableEntry,
+    BridgeTableRequest,
+    BridgeTableResponse,
+)
+from .join import (
+    JoinWithBridgeRequest,
+    JoinResponse,
+)
+
+__all__ = [
+    # Bridge models
+    "BridgeTableEntry",
+    "BridgeTableRequest",
+    "BridgeTableResponse",
+    # Join models
+    "JoinWithBridgeRequest",
+    "JoinResponse",
+]
+
